feat(component_interface_tools): add service log checker
