+++ common/component_interface_utils/CMakeLists.txt
@@ -3,5 +3,4 @@ project(component_interface_utils)
 
 find_package(autoware_cmake REQUIRED)
 autoware_package()
-ament_auto_add_executable(service_log_checker src/service_log_checker.cpp)
 ament_auto_package()
