--- common/component_interface_utils/src/service_log_checker.cpp
@@ -1,39 +1,0 @@
-// Copyright 2022 TIER IV, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#include "service_log_checker.hpp"
-
-#include <memory>
-
-ServiceLogChecker::ServiceLogChecker() : Node("service_log_checker")
-{
-  sub_ = create_subscription<ServiceLog>(
-    "/service_log", 50, std::bind(&ServiceLogChecker::on_service_log, this, std::placeholders::_1));
-}
-
-void ServiceLogChecker::on_service_log(const ServiceLog::ConstSharedPtr msg)
-{
-  RCLCPP_INFO_STREAM(get_logger(), msg->name);
-}
-
-int main(int argc, char ** argv)
-{
-  rclcpp::init(argc, argv);
-  rclcpp::executors::SingleThreadedExecutor executor;
-  auto node = std::make_shared<ServiceLogChecker>();
-  executor.add_node(node);
-  executor.spin();
-  executor.remove_node(node);
-  rclcpp::shutdown();
-}
